pyproject.toml
[project]
name = "dapple"
version = "0.1.0"
description = "Efficient and elegant plotting "
authors = [
    {name = "Daniel C. Jones"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cmap>=0.6.2",
    "freetype-py>=2.5.1",
    "matplotlib>=3.10.5",
    "moderngl>=5.12.0",
    "pillow>=11.3.0",
    "sympy>=1.14.0",
    "uharfbuzz>=0.51.2",
]

[build-system]
requires = ["uv_build>=0.8.10,<0.9.0"]
build-backend = "uv_build"

[tool.pyright]
venvPath = "."
venv = ".venv"

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]
